preparations for building x64 semantic server
git-svn-id: http://svn.dsource.org/projects/visuald/trunk@222 a09178c5-0f50-412e-aa4a-0aec1a98dccb

+++ vdc/ast/aggr.d
@@ -135,7 +135,7 @@ class Aggregate : Type
 	}
 
 	// always tree internal references
-	int mapDeclOffset;
+	size_t mapDeclOffset;
 	size_t[Declarator] mapDecl2Value;
 	size_t[string] mapName2Value;
 	Declarator[string] mapName2Method;
